+++ workers/google_import_queue_consumer/consumer.tsx
@@ -1,60 +1,31 @@
-import WorkEmailValidationEmail from "../../emails/invite-email";
-import * as React from "react";
-import {
-  EmailMessageType,
-  sendTransactionalEmail,
-} from "../../src/datasources/queues/mail";
-import { render } from "@react-email/render";
+import { createClient } from "@sanity/client";
+import { ensureKeys } from "../utils";
 
 type ENV = {
-  RESEND_EMAIL_KEY?: string;
+  SANITY_PROJECT_ID: string;
+  SANITY_DATASET: string;
+  SANITY_API_VERSION: string;
+  SANITY_SECRET_TOKEN: string;
 };
+
+const getSanityClient = (env: ENV) =>
+  createClient({
+    projectId: env.SANITY_PROJECT_ID,
+    dataset: env.SANITY_DATASET,
+    useCdn: false,
+    apiVersion: env.SANITY_API_VERSION, // use current date (YYYY-MM-DD) to target the latest API version
+    token: process.env.SANITY_SECRET_TOKEN, // Only if you want to update content with the client
+  });
 
 export const queueConsumer: ExportedHandlerQueueHandler<ENV> = async (
   batch,
   env,
 ) => {
+  ensureKeys(env);
+  const sanityClient = getSanityClient(env);
+  console.log("Processing queue", batch.queue);
   for await (const msg of batch.messages) {
-    switch (batch.queue) {
-      case "mail-queue-staging":
-      case "mail-queue-production":
-        await processEmailQueue(msg as Message<EmailMessageType>, env);
-        break;
-      default:
-        throw new Error(`Unknown queue ${batch.queue}`);
-    }
-  }
-};
-
-const processEmailQueue = async (
-  message: Message<EmailMessageType>,
-  env: ENV,
-) => {
-  try {
-    const { RESEND_EMAIL_KEY } = env;
-    if (!RESEND_EMAIL_KEY) {
-      throw new Error("RESEND_EMAIL_KEY is not defined");
-    }
-    // TODO: Send azure email
-    await sendTransactionalEmail(
-      {
-        RESEND_EMAIL_KEY,
-      },
-      {
-        from: "Javascript Chile <team@jschile.org>",
-        to: message.body.to,
-        subject: "Tu código de verificación",
-        html: render(
-          <WorkEmailValidationEmail
-            baseUrl=""
-            code={message.body.code}
-            userId={message.body.userId}
-          />,
-        ),
-      },
-    );
-    message.ack();
-  } catch (e) {
-    message.retry();
+    sanityClient.assets.upload("image", msg.data);
+    // console.log("Processing message", msg);
   }
 };
